Another minor folder tweak
Keeping class and instance separate means I don't accidentally call e.g. the imported `logger` instead of `this.logger`
