+++ Dockerfile
@@ -4,4 +4,4 @@ COPY package*.json ./
 RUN npm ci
 COPY . .
 RUN npm run build
-CMD ["node", "dist/commandHandler.js"]
+CMD ["node", "dist/index.js"]
