+++ src/website/interface.ts
@@ -1,0 +1,38 @@
+export interface WebsiteWrapper {
+	createTournament(name: string, desc: string): Promise<WebsiteTournament>;
+	updateTournament(tournamentId: string, name: string, desc: string): Promise<void>;
+	getTournament(tournamentId: string): Promise<WebsiteTournament>;
+	registerPlayer(tournamentId: string, playerName: string, playerId: string): Promise<number>;
+}
+
+// interface structure WIP as fleshed out command-by-command
+export interface WebsiteTournament {
+	id: string;
+	name: string;
+	desc: string;
+	url: string;
+	players: number[];
+}
+
+export class WebsiteInterface {
+	private api: WebsiteWrapper;
+	constructor(api: WebsiteWrapper) {
+		this.api = api;
+	}
+
+	public async createTournament(name: string, desc: string): Promise<WebsiteTournament> {
+		return await this.api.createTournament(name, desc);
+	}
+
+	public async updateTournament(tournamentId: string, name: string, desc: string): Promise<void> {
+		return await this.api.updateTournament(tournamentId, name, desc);
+	}
+
+	public async getTournament(tournamentId: string): Promise<WebsiteTournament> {
+		return await this.api.getTournament(tournamentId);
+	}
+
+	public async registerPlayer(tournamentId: string, playerName: string, playerId: string): Promise<number> {
+		return await this.api.registerPlayer(tournamentId, playerName, playerId);
+	}
+}
